Moved job models to its own namespace
